requirements.txt
Brotli==1.1.0
cffi==1.16.0
chardet==5.2.0
cssselect2==0.7.0
fonttools==4.50.0
freetype-py==2.3.0
html5lib==1.1
pillow==10.3.0
pycairo==1.26.0
pycparser==2.22
pydyf==0.9.0
pyphen==0.14.0
python-dotenv==1.0.1
reportlab==4.1.0
rlPyCairo==0.3.0
six==1.16.0
tinycss2==1.2.1
weasyprint==61.2
webencodings==0.5.1
zopfli==0.2.3
